requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.3
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
proxy_tools==0.1.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywebview==5.4
setuptools==80.8.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
